config.py
import os
from dotenv import load_dotenv

load_dotenv()

basedir = os.path.abspath(os.path.dirname(__file__))


class Config:
    SECRET_KEY = os.getenv("SECRET_KEY")
    SQLALCHEMY_DATABASE_URI = os.getenv("DATABASE_URI") or "sqlite:///" + os.path.join(
        basedir, "app.db"
    )
    SQLALCHEMY_TRACK_MODIFICATIONS = False
    FLASK_ADMIN_SWATCH = "cerulean"
    CLOUDINARY_CLOUD_NAME = os.getenv("CLOUD_NAME")
    CLOUDINARY_API_KEY = os.getenv("API_KEY")
    CLOUDINARY_API_SECRET = os.getenv("API_SECRET")
