requirements.txt
alembic==1.11.1
blinker==1.6.2
certifi==2023.7.22
click==8.1.6
cloudinary==1.33.0
Flask==2.3.2
Flask-Admin==1.6.1
Flask-Login==0.6.2
Flask-Migrate==4.0.4
Flask-SQLAlchemy==3.0.5
greenlet==2.0.2
gunicorn==21.2.0
itsdangerous==2.1.2
Jinja2==3.1.2
Mako==1.2.4
MarkupSafe==2.1.3
numpy==1.25.2
packaging==23.1
Pillow==10.0.0
psycopg2-binary==2.9.6
python-dotenv==1.0.0
six==1.16.0
SQLAlchemy==2.0.19
typing_extensions==4.7.1
urllib3==1.26.16
Werkzeug==2.3.6
WTForms==3.0.1
